requirements.txt
Django==2.0.4
django-bootstrap3==9.1.0
django-filter==1.1.0
mysqlclient==1.3.12
pytz==2018.4
django-braces==1.13.0
django-extra-views
django-query-parameters==0.2.3
django-extensions